% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocbe.R
\docType{package}
\name{ocbe}
\alias{ocbe}
\title{Misc functions from OCBE/UiO and OCBE/OUS}
\description{
This package contains miscellaneous R functions created by 
 staff members of the Oslo Centre for Biostatistics and Epidemiology (OCBE).
 OCBE is a joint center between the University of Oslo and the Oslo 
 University Hospital.
}
\section{Core functions}{

 - No functions implemented yet.
}

